+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make `Unit` object-safe.
 - Allow comments in certain locations in the assembly.
 - Make `ty` module visible.
+- Change `const` instructions to use `IntValue` and `TimeValue`.
 
 ## 0.9.0 - 2019-10-24
 ### Added